requirements.txt
twine
build
setuptools
pydantic_core
awswrangler